Extract AbstractJsonSchemaGenerator to remove duplication between JsonSchemaGenerator and McpJsonSchemaGenerator
